third_party/WebKit/LayoutTests/editing/selection/modify_extend/extend_forward_line_crash.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
    assert_selection(
        `<div>|foo<object style="content: counters(c, &quot;.&quot;) &quot;-&quot;" type="text/vbs"></object></div>`,
        selection => selection.modify('extend', 'forward', 'line'),
        `<div>^foo|<object style="content: counters(c, &quot;.&quot;) &quot;-&quot;" type="text/vbs"></object></div>`);
}, 'exnted forward line should not crash with OBJECT element');
</script>
